Added title to notes

+++ lib/views/notes/create_update_note_view.dart
@@ -13,6 +13,7 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
   DatabaseNote? _note;
   late final NotesService _notesService;
   late final TextEditingController _textController;
+  late final TextEditingController _titleController;
 
   Future<DatabaseNote> createNewNote() async {
     final existingNote = _note;
@@ -28,7 +29,9 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
   //if text is empty, delete note
   void _deleteNote() {
     final note = _note;
-    if (_textController.text.isEmpty && note != null) {
+    if (_titleController.text.isEmpty &&
+        _textController.text.isEmpty &&
+        note != null) {
       _notesService.deleteNote(id: note.id);
     }
   }
@@ -37,9 +40,11 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
   void _saveNote() async {
     final note = _note;
     final text = _textController.text;
-    if (note != null && text.isNotEmpty) {
+    final title = _titleController.text;
+    if (note != null && title.isNotEmpty && text.isNotEmpty) {
       await _notesService.updateNote(
         note: note,
+        title: title,
         text: text,
       );
     }
@@ -51,9 +56,27 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
       return;
     }
     final text = _textController.text;
+    final title = _titleController.text;
+
     await _notesService.updateNote(
       note: note,
       text: text,
+      title: title,
+    );
+  }
+
+  void _titleControllerListener() async {
+    final note = _note;
+    if (note == null) {
+      return;
+    }
+    final text = _textController.text;
+    final title = _titleController.text;
+
+    await _notesService.updateNote(
+      note: note,
+      text: text,
+      title: title,
     );
   }
 
@@ -62,10 +85,29 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
     _textController.addListener(_textControllerListener);
   }
 
+  void _setupTitleControllerListener() {
+    _titleController.removeListener(_titleControllerListener);
+    _titleController.addListener(_titleControllerListener);
+  }
+
+  Future<void> _createNoteAndSetupControllers() async {
+    final note = await createNewNote();
+    setState(() {
+      _note = note;
+    });
+
+    _setupTextControllerListener();
+    _setupTitleControllerListener();
+  }
+
   @override
   void initState() {
     _notesService = NotesService();
     _textController = TextEditingController();
+    _titleController = TextEditingController();
+
+    _createNoteAndSetupControllers();
+
     super.initState();
   }
 
@@ -74,6 +116,7 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
     _deleteNote();
     _saveNote();
     _textController.dispose();
+    _titleController.dispose();
     super.dispose();
   }
 
@@ -81,6 +124,7 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
+        backgroundColor: const Color(0xFF7EABFF),
         title: const Text('New Note'),
       ),
       body: FutureBuilder(
@@ -89,13 +133,26 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
           switch (snapshot.connectionState) {
             case ConnectionState.done:
               _note = snapshot.data as DatabaseNote;
-              _setupTextControllerListener();
-              return TextField(
-                controller: _textController,
-                keyboardType: TextInputType.multiline,
-                maxLines: null,
-                decoration: const InputDecoration(
-                    hintText: 'Start typing your note here'),
+              return Column(
+                children: [
+                  Padding(
+                    padding: const EdgeInsets.only(
+                        left: 10.0, right: 10.0, top: 20.0, bottom: 10.0),
+                    child: TextField(
+                      controller: _titleController,
+                      keyboardType: TextInputType.text,
+                      maxLines: null,
+                      decoration: const InputDecoration(hintText: 'Title'),
+                    ),
+                  ),
+                  TextField(
+                    controller: _textController,
+                    keyboardType: TextInputType.multiline,
+                    maxLines: null,
+                    decoration: const InputDecoration(
+                        hintText: 'Start typing your note here'),
+                  ),
+                ],
               );
             default:
               return const CircularProgressIndicator();
